+++ .github/ISSUE_TEMPLATE/issue-template.md
@@ -1,0 +1,17 @@
+---
+name: Issue template
+about: 기능 추가 이슈 템플릿
+title: feat
+labels: ""
+assignees: Jake1152
+---
+
+## 설명
+
+ex) quiz 결과 저장 api를 구현합니다.
+
+## 할 일
+
+- [ ] body validator
+- [ ] transaction 처리
+- [ ] 다중 테스트
