+++ src/queries/quizQuery.js
@@ -1,0 +1,1 @@
+exports.quizRequest = `SELECT word, definetion, initial_constants FROM quiz`;
