+++ src/queries/scoreQuery.js
@@ -1,0 +1,2 @@
+exports.getScoreInfo = `SELECT * FROM score WHERE user_id = ?`;
+// exports.saveScoreInfo = `UPDATE score SET total_scores, total`;
